# app/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "FizzBuzz",
        "title": "FizzBuzz Swagger UI",
        "termsOfService": "demo.com",
        "contact": {
            "name": "API Support",
            "url": "http://demo.com/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "description": "Returns a JSON response with FizzBuzz result",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fizzbuzz"
                ],
                "summary": "FizzBuzz",
                "operationId": "fizzbuzz",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Int1",
                        "name": "int1",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Int2",
                        "name": "int2",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Str1",
                        "name": "str1",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Str2",
                        "name": "str2",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/fizzbuzz_pkg_apiresponse.ApiResponse-string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/fizzbuzz_pkg_apiresponse.ApiResponse-string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/fizzbuzz_pkg_apiresponse.ApiResponse-string"
                        }
                    }
                }
            }
        },
        "/stat": {
            "get": {
                "description": "Returns a JSON response of the query string with the most hits",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fizzbuzz"
                ],
                "summary": "Stat",
                "operationId": "fizzbuzz-stat",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fizzbuzz_pkg_apiresponse.ApiResponse-string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/fizzbuzz_pkg_apiresponse.ApiResponse-string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "fizzbuzz_pkg_apiresponse.ApiResponse-string": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        }
    }
}